requirements.txt
kafka-python
psycopg2
pytest
pytest-cov